+++ lib/src/lunar_calendar/models/two_hour_period.dart
@@ -13,28 +13,38 @@ class TwoHourPeriod extends Equatable {
   /// 是否吉时
   final bool isLucky;
 
-  const TwoHourPeriod(
-      {required this.nameInTheDay,
-      required this.twoHourPeriodIndex,
-      required this.isLucky});
+  /// 刻：1时辰分8刻，包括上四刻和下四刻
+  final Ke ke;
+
+  const TwoHourPeriod({
+    required this.nameInTheDay,
+    required this.twoHourPeriodIndex,
+    required this.isLucky,
+    required this.ke,
+  });
 
   factory TwoHourPeriod.from({required DateTime dateTime}) {
     final twoHourPeriodIndex = getTwoHourPeriodIndex(dateTime.hour);
     final nameInTheDay = getTwoHour8Char(dateTime: dateTime);
     final isLucky = getTwoHourPeriodLuckyList(dateTime)[twoHourPeriodIndex];
     return TwoHourPeriod(
-        nameInTheDay: nameInTheDay,
-        twoHourPeriodIndex: twoHourPeriodIndex,
-        isLucky: isLucky);
+      nameInTheDay: nameInTheDay,
+      twoHourPeriodIndex: twoHourPeriodIndex,
+      isLucky: isLucky,
+      ke: Ke.from(dateTime: dateTime),
+    );
   }
 
   @override
-  List<Object?> get props => [nameInTheDay, twoHourPeriodIndex, isLucky];
+  List<Object?> get props => [nameInTheDay, twoHourPeriodIndex, isLucky, ke];
 }
 
 extension TwoHourPeriodX on TwoHourPeriod {
   /// 地支名
-  String get name => the12EarthlyBranches[twoHourPeriodIndex];
+  String get name => nameInTheDay[1];
+
+  /// 天干名
+  String get steamName => nameInTheDay[0];
 
   /// 时辰汉代名
   String get nameInHanDynasty => twoHourPeriodHanName[twoHourPeriodIndex];